Move nibe_heatpump base entity to separate module
