+++ homeassistant/components/nibe_heatpump/coordinator.py
@@ -17,12 +17,7 @@ from nibe.heatpump import HeatPump, Series
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import CALLBACK_TYPE, HomeAssistant, callback
 from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.entity import async_generate_entity_id
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-    UpdateFailed,
-)
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import DOMAIN, LOGGER
 
@@ -68,7 +63,7 @@ class ContextCoordinator[_DataTypeT, _ContextTypeT](DataUpdateCoordinator[_DataT
         return release_update
 
 
-class Coordinator(ContextCoordinator[dict[int, CoilData], int]):
+class CoilCoordinator(ContextCoordinator[dict[int, CoilData], int]):
     """Update coordinator for nibe heat pumps."""
 
     config_entry: ConfigEntry
@@ -188,43 +183,3 @@ class Coordinator(ContextCoordinator[dict[int, CoilData], int]):
             self.task.cancel()
             await asyncio.wait((self.task,))
         await self.connection.stop()
-
-
-class CoilEntity(CoordinatorEntity[Coordinator]):
-    """Base for coil based entities."""
-
-    _attr_has_entity_name = True
-    _attr_entity_registry_enabled_default = False
-
-    def __init__(
-        self, coordinator: Coordinator, coil: Coil, entity_format: str
-    ) -> None:
-        """Initialize base entity."""
-        super().__init__(coordinator, {coil.address})
-        self.entity_id = async_generate_entity_id(
-            entity_format, coil.name, hass=coordinator.hass
-        )
-        self._attr_name = coil.title
-        self._attr_unique_id = f"{coordinator.unique_id}-{coil.address}"
-        self._attr_device_info = coordinator.device_info
-        self._coil = coil
-
-    @property
-    def available(self) -> bool:
-        """Return if entity is available."""
-        return self.coordinator.last_update_success and self._coil.address in (
-            self.coordinator.data or {}
-        )
-
-    def _async_read_coil(self, data: CoilData):
-        """Update state of entity based on coil data."""
-
-    async def _async_write_coil(self, value: float | str):
-        """Write coil and update state."""
-        await self.coordinator.async_write_coil(self._coil, value)
-
-    def _handle_coordinator_update(self) -> None:
-        data = self.coordinator.data.get(self._coil.address)
-        if data is not None:
-            self._async_read_coil(data)
-        self.async_write_ha_state()
