+++ homeassistant/components/nibe_heatpump/water_heater.py
@@ -26,7 +26,7 @@ from .const import (
     VALUES_TEMPORARY_LUX_INACTIVE,
     VALUES_TEMPORARY_LUX_ONE_TIME_INCREASE,
 )
-from .coordinator import Coordinator
+from .coordinator import CoilCoordinator
 
 
 async def async_setup_entry(
@@ -36,7 +36,7 @@ async def async_setup_entry(
 ) -> None:
     """Set up platform."""
 
-    coordinator: Coordinator = hass.data[DOMAIN][config_entry.entry_id]
+    coordinator: CoilCoordinator = hass.data[DOMAIN][config_entry.entry_id]
 
     def water_heaters():
         for key, group in WATER_HEATER_COILGROUPS.get(coordinator.series, ()).items():
@@ -48,7 +48,7 @@ async def async_setup_entry(
     async_add_entities(water_heaters())
 
 
-class WaterHeater(CoordinatorEntity[Coordinator], WaterHeaterEntity):
+class WaterHeater(CoordinatorEntity[CoilCoordinator], WaterHeaterEntity):
     """Sensor entity."""
 
     _attr_entity_category = None
@@ -59,7 +59,7 @@ class WaterHeater(CoordinatorEntity[Coordinator], WaterHeaterEntity):
 
     def __init__(
         self,
-        coordinator: Coordinator,
+        coordinator: CoilCoordinator,
         key: str,
         desc: WaterHeaterCoilGroup,
     ) -> None:
